Optimization for the scheduler sampler BPF. Replaces the hashmap for PIDs with an array to reduce lookup time at the expense of requiring additional memory. This raises the memory overhead for the BPF sampler by 64MB.

On my test system, running a compute-heavy workload. This reduces the overheads on scheduler switch and scheduler wakeup.
switch: 542ns -> 223ns
wakeup: 300ns -> 165ns